Move word.ts to /common
