+++ extensions/ql-vscode/src/common/word.ts
