+++ extensions/ql-vscode/src/query-history/history-item-label-provider.ts
@@ -9,7 +9,7 @@ import {
 } from "./query-history-info";
 import { VariantAnalysisHistoryItem } from "./variant-analysis-history-item";
 import { assertNever } from "../common/helpers-pure";
-import { pluralize } from "../pure/word";
+import { pluralize } from "../common/word";
 import { humanizeQueryStatus } from "../query-status";
 
 interface InterpolateReplacements {
