+++ extensions/ql-vscode/src/query-history/query-history-info.ts
@@ -1,7 +1,7 @@
 import { VariantAnalysisHistoryItem } from "./variant-analysis-history-item";
 import { LocalQueryInfo } from "../query-results";
 import { assertNever } from "../common/helpers-pure";
-import { pluralize } from "../pure/word";
+import { pluralize } from "../common/word";
 import {
   hasRepoScanCompleted,
   getActionsWorkflowRunUrl as getVariantAnalysisActionsWorkflowRunUrl,
